feat: add _permission() methods for category
